requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.0.5
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.14.0
fonttools==4.37.2
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
hydra-core==1.3.2
idna==3.4
importlib_resources==6.4.0
Jinja2==3.1.4
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
omegaconf==2.3.0
packaging==21.3
pillow==10.3.0
piq==0.8.0
platformdirs==4.2.2
prettytable==3.10.0
protobuf==4.25.3
psutil==5.9.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
scipy==1.10.1
sentry-sdk==2.2.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tqdm==4.64.1
triton==2.3.0
typing_extensions==4.11.0
urllib3==1.26.12
wandb==0.17.0
wcwidth==0.2.13
imageio==2.34.2
imageio-ffmpeg==0.5.1
torchdiffeq==0.2.5
diffusers==0.32.2
huggingface-hub==0.29.3
importlib-metadata==8.5.0
regex-2024==11.6
safetensors==0.5.3
zipp==3.20.2
tokenizers==0.20.3
transformers==4.46.3
pytorch-lightning==1.0.8
einops==0.8.1
kornia==0.7.3
-e git+https://github.com/openai/CLIP.git@main#egg=clip